backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
stripe==7.4.0
django-cors-headers==4.3.1
drf-spectacular==0.26.5
gunicorn==21.2.0
python-dateutil==2.8.2
Pillow==10.1.0
